--- SpaceBattle.Lib/ClassForTests.cs
@@ -1,12 +1,0 @@
-﻿public class ClassForTests
-{
-    public readonly int Position;
-    public readonly IRotateble Rotable;
-    public readonly string Name;
-    public ClassForTests(IRotateble Rotable, int Position, string Name)
-    {
-        this.Rotable = Rotable;
-        this.Position = Position;
-        this.Name = Name;
-    }
-}
