+++ SpaceBattle.Tests/ClassForTests.cs
@@ -1,0 +1,12 @@
+﻿public class ClassForTests
+{
+    public readonly string Name;
+    public readonly int Count;
+    public readonly IRotateble Rotateble;
+    public ClassForTests(string Name, int Count, IRotateble Rotateble)
+    {
+        this.Name = Name;
+        this.Count = Count;
+        this.Rotateble = Rotateble;
+    }
+}
